requirements.txt
requests
beautifulsoup4
urllib3<=1.26
python-dotenv
google-cloud-webrisk